Add share popup to mobile view of app details
